requirements.txt
streamlit
pandas
plotly
google-generativeai
stripe
phonenumbers
pytz
openpyxl